Revert "update kernel config for 4.6.4"
The default manifest should not track a specific revision
of the kernel configurations. It brings zero benefit, since
reproductibility is pointless for the default conguration.
It also would cause a lot more of work because one would need
to make two commits instead of just one.

It makes a lot of sense for specific defconfigs, but not
for the default one.

This reverts commit fe2f304fe35e013c48e982ce1b62ce1c5584aa0c.